proyecto-melendez-master/server/requirements.txt
aiofiles==23.2.1
certifi==2023.7.22
dnspython==2.4.2
html5tagger==1.3.0
httptools==0.6.0
multidict==6.0.4
orjson==3.9.5
packaging==23.1
pymongo==4.5.0
python-dotenv==1.0.0
sanic==23.6.0
Sanic-Cors==2.2.0
sanic-routing==23.6.0
tracerite==1.1.0
typing_extensions==4.7.1
ujson==5.8.0
uvloop==0.17.0
websockets==11.0.3
